Extract common path helper methods to PathHelper class
